fix param type from double to float. (#537)
* fix api comment

* fix api comment

* fix api comment

* change param name in gis api

* change param name in gis api

* change scheam name

* change scheam name

* fix api comment

* ...

* fix Copyright

* fix api comment

* fix api comment

* fix gis api comment

* fix api comment

* modified param name

* modify plot api comment

* modify plot api comment

* modify plot api comment

* add register_func comment

* modify register_func comment

* change import order in plot test

* fix param type from double to float

+++ python/arctern/_wrapper_func.py
@@ -531,7 +531,7 @@ def ST_SimplifyPreserveTopology(geos, distance_tolerance):
     :type geos: Series(dtype: object)
     :param geos: Geometries in WKB from.
 
-    :type distance_tolerance: double
+    :type distance_tolerance: float
     :param distance_tolerance: The maximum distance between a point on a linestring and a curve.
 
     :rtype: Series(dtype: object)
@@ -971,7 +971,7 @@ def ST_Buffer(geos, distance):
     :type geos: Series(dtype: object)
     :param geos: Geometries in WKB form.
 
-    :type distance: double
+    :type distance: float
     :param distance: The maximum distance of the returned geometry from the given geometry.
 
     :rtype: Series(dtype: object)
